refactor: getPlayerResult 메서드 이관
Dealer -> GameResult
