+++ src/main/kotlin/blackjack/domain/GameResult.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 class DealerResult() {
     var win: Int = 0
@@ -38,9 +38,20 @@ class GameResult(private val dealer: Dealer, val players: List<Player>) {
         }
     }
 
+    fun getPlayerResult(player: Player): GameResultStatus {
+        if (player.isBust()) return GameResultStatus.PLAYER_LOSE
+        if (dealer.isBust()) return GameResultStatus.PLAYER_WIN
+        return when {
+            dealer.totalSum > player.totalSum -> GameResultStatus.PLAYER_LOSE
+            player.totalSum > dealer.totalSum -> GameResultStatus.PLAYER_WIN
+            player.totalSum == dealer.totalSum -> GameResultStatus.DRAW
+            else -> throw IllegalArgumentException()
+        }
+    }
+
     fun getResult(): GameResult {
         players.forEach { player ->
-            val playerResult = dealer.getPlayerResult(player)
+            val playerResult = getPlayerResult(player)
             updateResult(player, playerResult)
         }
         return this
