+++ src/main/kotlin/blackjack/domain/GameResultStatus.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 enum class GameResultStatus {
     PLAYER_WIN,
