+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -1,16 +1,16 @@
 package blackjack.view
 
-import blackjack.Dealer
-import blackjack.GameResult
-import blackjack.Player
-import blackjack.toDisplayName
+import blackjack.domain.Dealer
+import blackjack.domain.GameResult
+import blackjack.domain.Player
+import blackjack.domain.toDisplayName
 
 object OutputView {
     fun showInitialCards(
         dealer: Dealer,
         players: List<Player>,
     ) {
-        println("딜러와 ${players.joinToString(", ") { it.name }}에게 2장을 나누었습니다.\n")
+        println("딜러와 ${players.joinToString { it.name }}에게 2장을 나누었습니다.\n")
         println("딜러: ${dealer.cards.first()}")
         players.forEach { player ->
             printPlayerCards(player)
@@ -18,7 +18,7 @@ object OutputView {
     }
 
     fun printPlayerCards(player: Player) {
-        println("${player.name}카드: ${player.cards.joinToString(", ")}")
+        println("${player.name}카드: ${player.cards.joinToString()}")
     }
 
     fun printDealerHaveAdditionalCard() {
@@ -29,16 +29,16 @@ object OutputView {
         dealer: Dealer,
         players: List<Player>,
     ) {
-        println("딜러 카드: ${dealer.cards.joinToString(", ")} - 결과: ${dealer.totalSum}")
+        println("딜러 카드: ${dealer.cards.joinToString()} - 결과: ${dealer.totalSum}")
 
         players.forEach { player ->
-            println("${player.name}카드: ${player.cards.joinToString(", ")} - 결과: ${player.totalSum}")
+            println("${player.name}카드: ${player.cards.joinToString()} - 결과: ${player.totalSum}")
         }
     }
 
     fun printGameResult(result: GameResult) {
         println("\n##최종 승패")
-        println("딜러: ${result.dealerResult.win}승 ${result.dealerResult.lose}패")
+        println("딜러: ${result.dealerResult.win}승 ${result.dealerResult.lose}패 ${result.dealerResult.draw}무")
 
         result.playerResultStatus.entries.forEach { it ->
             println("${it.key.name}: ${it.value.toDisplayName()} ")
