+++ src/test/kotlin/blackjack/FakeParticipant.kt
@@ -1,5 +1,7 @@
 package blackjack
 
+import blackjack.domain.Participant
+
 class FakeParticipant : Participant() {
     override val hitThreshold: Int
         get() = 21
